change all methods to f64

+++ src/filter.rs
@@ -10,7 +10,7 @@ use ndarray::prelude::*;
 /// * Replicate: copy the value at the border, i.e., aaaa|abcdefgh|hhhh
 #[derive(Copy, Clone)]
 pub enum BorderType {
-    Constant(u8),
+    Constant(f64),
     Reflect,
     Replicate
 }
@@ -50,7 +50,7 @@ pub fn border_interpolate(p:i32, len:usize, border: BorderType) -> Option<usize>
 /// Get the value of an image at a location which may be outside the image.
 ///
 /// This function is used by `filter`.
-fn access_img_border(src: &Array<u8, Ix2>, x:i32, y:i32, border: BorderType) -> u8 {
+fn access_img_border(src: &Array<f64, Ix2>, x:i32, y:i32, border: BorderType) -> f64 {
     if x >= 0 && y >= 0 && x < src.shape()[0] as i32 && y < src.shape()[1] as i32 {
         src[[x as usize, y as usize]]
     }
@@ -74,8 +74,8 @@ fn access_img_border(src: &Array<u8, Ix2>, x:i32, y:i32, border: BorderType) -> 
 /// The method of dealing with border situation is selected by `border`. By setting
 /// `border=Reflect`, you will get the default result of OpenCV.
 ///
-pub fn filter_(src: &Array<u8, Ix2>, kernel: &Array<f64, Ix2>, border: BorderType, 
-               out:&mut Array<u8, Ix2>) {
+pub fn filter_(src: &Array<f64, Ix2>, kernel: &Array<f64, Ix2>, border: BorderType, 
+               out:&mut Array<f64, Ix2>) {
     let kh = kernel.shape()[0];
     let kw = kernel.shape()[1];
     let kcx = (kh / 2) as i32; // kernel center x
@@ -89,11 +89,11 @@ pub fn filter_(src: &Array<u8, Ix2>, kernel: &Array<f64, Ix2>, border: BorderTyp
                 for kj in 0..kw {
                     let sx = i as i32 + ki as i32 - kcx;
                     let sy = j as i32 + kj as i32 - kcy;
-                    let sval = access_img_border(src, sx, sy, border) as f64;
+                    let sval = access_img_border(src, sx, sy, border);
                     val = val + sval * kernel[[ki, kj]];
                 }
             }
-            out[[i, j]] = val as u8;
+            out[[i, j]] = val;
         }
     }
 }
@@ -107,24 +107,24 @@ pub fn filter_(src: &Array<u8, Ix2>, kernel: &Array<f64, Ix2>, border: BorderTyp
 /// ```
 ///  use simplecv::filter::*;
 ///  use ndarray::arr2;
-///  let A = arr2(&[[0, 0, 0, 0, 0],
-///                 [0, 1, 1, 1, 0],
-///                 [0, 1, 1, 1, 0],
-///                 [0, 1, 1, 1, 0],
-///                 [0, 0, 0, 0, 0]]);
+///  let A = arr2(&[[0.0, 0.0, 0.0, 0.0, 0.0],
+///                 [0.0, 1.0, 1.0, 1.0, 0.0],
+///                 [0.0, 1.0, 1.0, 1.0, 0.0],
+///                 [0.0, 1.0, 1.0, 1.0, 0.0],
+///                 [0.0, 0.0, 0.0, 0.0, 0.0]]);
 ///  let kernel = arr2(&[[1.0, 2.0, 1.0],
 ///                      [2.0, 4.0, 2.0],
 ///                      [1.0, 2.0, 1.0]]);
-///  let target = arr2(&[[1, 3, 4, 3, 1],
-///                      [3, 9, 12, 9, 3],
-///                      [4, 12, 16, 12, 4],
-///                      [3, 9, 12, 9, 3],
-///                      [1, 3, 4, 3, 1]]);
-///  let output = filter(&A, &kernel, BorderType::Constant(0));
+///  let target = arr2(&[[1.0, 3.0, 4.0, 3.0, 1.0],
+///                      [3.0, 9.0, 12.0, 9.0, 3.0],
+///                      [4.0, 12.0, 16.0, 12.0, 4.0],
+///                      [3.0, 9.0, 12.0, 9.0, 3.0],
+///                      [1.0, 3.0, 4.0, 3.0, 1.0]]);
+///  let output = filter(&A, &kernel, BorderType::Constant(0.0));
 ///  assert_eq!(target, output);
 /// ```
 ///
-pub fn filter(src: &Array<u8, Ix2>, kernel: &Array<f64, Ix2>, border: BorderType) -> Array<u8, Ix2>{
+pub fn filter(src: &Array<f64, Ix2>, kernel: &Array<f64, Ix2>, border: BorderType) -> Array<f64, Ix2>{
     let shape = src.shape();
     let height = shape[0] as usize;
     let width = shape[1] as usize;
